+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 - new MINOR version for added functionality in a backwards compatible manner
 - new PATCH version for backwards compatible bug fixes
 
+v2.1.7
+---------
+2023-10-18:
+    - validator for "Criteria.ids" and "Criteria.limit"
+      if "Criteria.ids" are passed, set limits to the number of ids and page to 1
+
 v2.1.6
 ---------
 2023-10-18: